backend/requirements.txt
Django>=6.0,<7
djangorestframework>=3.15
drf-spectacular>=0.27
django-cors-headers>=4.6
